LayoutTests/http/tests/security/no-indexeddb-from-sandbox-expected.txt

CONSOLE MESSAGE: PASS: db.open() threw a SECURITY_ERR!
